+++ CHANGELOG.md
@@ -10,6 +10,7 @@ SPDX-License-Identifier: CC0-1.0 -->
 - Add code formatting using spotless #117 (@dlr-cjs, @dlr_fn, @dlr_jk)
 - Add new class `GenericDevice` that represents a generic electrical flexibility #116 (@dlr-cjs, @dlr_fn, @dlr_jk)
 - Add optional "Refinancing" parameters for `GreenHydrogenTrader` #149 (@dlr_jk)
+- Add JSONable interface for UrlModelService input classes #151 (@dlr_elghazi, @dlr-cjs)
 
 ### Fixed
 - ReadMe: Fixed typo in Acknowledgements (@dlr_fn)